+++ src/dk/aau/cs/io/TapnLegacyXmlLoader.java
@@ -675,8 +675,8 @@ public class TapnLegacyXmlLoader {
 						String arcTempX = element.getAttribute("x");
 						String arcTempY = element.getAttribute("y");
 						String arcTempType = element.getAttribute("curvePoint");
-						float arcPointX = Float.valueOf(arcTempX);
-						float arcPointY = Float.valueOf(arcTempY);
+						double arcPointX = Double.valueOf(arcTempX);
+						double arcPointY = Double.valueOf(arcTempY);
 						arcPointX += Pipe.ARC_CONTROL_POINT_CONSTANT + 1;
 						arcPointY += Pipe.ARC_CONTROL_POINT_CONSTANT + 1;
 						boolean arcPointType = Boolean.valueOf(arcTempType);
