+++ src/pipe/gui/graphicElements/ArcPath.java
@@ -34,10 +34,8 @@ public class ArcPath implements Shape {
 	private List<ArcPathPoint> pathPoints = new ArrayList<ArcPathPoint>();
 	private Arc myArc;
 	private boolean pointLock = false;
-	private static Stroke proximityStroke = new BasicStroke(
-			Pipe.ARC_PATH_PROXIMITY_WIDTH);
-	private static Stroke stroke = new BasicStroke(
-			Pipe.ARC_PATH_SELECTION_WIDTH);
+	private static Stroke proximityStroke = new BasicStroke(Pipe.ARC_PATH_PROXIMITY_WIDTH);
+	private static Stroke stroke = new BasicStroke(Pipe.ARC_PATH_SELECTION_WIDTH);
 	private Shape shape, proximityShape;
 	private int transitionAngle;
 	private final static boolean showDebugCurvedControlPoints = false;
@@ -118,9 +116,8 @@ public class ArcPath implements Shape {
 
 		//Calculate the arc mid-point for lable placement
 		if (getEndIndex() < 2) {
-			midPoint.x = (float) (((pathPoints.get(0)).getPoint().x + (pathPoints
-					.get(1)).getPoint().x) * 0.5);
-			midPoint.y = (float) (((pathPoints.get(0)).getPoint().y + (pathPoints
+			midPoint.x = (((pathPoints.get(0)).getPoint().x + (pathPoints.get(1)).getPoint().x) * 0.5);
+			midPoint.y = (((pathPoints.get(0)).getPoint().y + (pathPoints
 					.get(1)).getPoint().y) * 0.5);
 		} else {
 			double acc = 0;
@@ -129,8 +126,7 @@ public class ArcPath implements Shape {
 				ArcPathPoint previousPoint = currentPoint;
 				currentPoint = pathPoints.get(c);
 
-				double inc = getMod(currentPoint.getPoint(), previousPoint
-						.getPoint());
+				double inc = getMod(currentPoint.getPoint(), previousPoint.getPoint());
 				if ((acc + inc > length)) {
 					percent = (length - acc) / inc;
 					break;
@@ -139,12 +135,9 @@ public class ArcPath implements Shape {
 			}
 
 			ArcPathPoint previousPoint = pathPoints.get(c - 1);
-			midPoint.x = previousPoint.getPoint().x
-					+ (float) ((currentPoint.getPoint().x - previousPoint
-							.getPoint().x) * percent);
+			midPoint.x = previousPoint.getPoint().x + ((currentPoint.getPoint().x - previousPoint.getPoint().x) * percent);
 			midPoint.y = previousPoint.getPoint().y
-					+ (float) ((currentPoint.getPoint().y - previousPoint
-							.getPoint().y) * percent);
+					+ ((currentPoint.getPoint().y - previousPoint.getPoint().y) * percent);
 		}
 
 		shape = stroke.createStrokedShape(this);
@@ -159,9 +152,13 @@ public class ArcPath implements Shape {
 	}
 
 	/* returns a control point for curve CD with incoming vector AB */
-	private Point2D.Float getControlPoint(Point2D.Float A, Point2D.Float B,
-			Point2D.Float C, Point2D.Float D) {
-		Point2D.Float p = new Point2D.Float(0, 0);
+	private Point2D.Double getControlPoint(
+	    Point2D.Double A,
+        Point2D.Double B,
+        Point2D.Double C,
+        Point2D.Double D
+    ) {
+		Point2D.Double p = new Point2D.Double(0, 0);
 
 		double modAB = getMod(A, B);
 		double modCD = getMod(C, D);
@@ -172,15 +169,15 @@ public class ArcPath implements Shape {
 		if (modAB < 7) {
 			// hack, stops division by zero, modAB can only be this low if the
 			// points are virtually superimposed anyway
-			p = (Point2D.Float) C.clone();
+			p = (Point2D.Double) C.clone();
 		} else {
-			p.x = C.x + (float) (ABx * modCD / Pipe.ARC_CONTROL_POINT_CONSTANT);
-			p.y = C.y + (float) (ABy * modCD / Pipe.ARC_CONTROL_POINT_CONSTANT);
+			p.x = C.x + (ABx * modCD / Pipe.ARC_CONTROL_POINT_CONSTANT);
+			p.y = C.y + (ABy * modCD / Pipe.ARC_CONTROL_POINT_CONSTANT);
 		}
 		return p;
 	}
 
-	private double getMod(Point2D.Float A, Point2D.Float B) {
+	private double getMod(Point2D.Double A, Point2D.Double B) {
 		double ABx = A.x - B.x;
 		double ABy = A.y - B.y;
 
@@ -233,10 +230,8 @@ public class ArcPath implements Shape {
 
 				for (int k2 = 1; k2 <= lengthOfCurve; k2++) {
 					myCurrentPoint = pathPoints.get(k2 + curveStartIndex);
-					myCurrentPoint.setControl1(X[k2 - 1].getX1(), Y[k2 - 1]
-							.getX1());
-					myCurrentPoint.setControl2(X[k2 - 1].getX2(), Y[k2 - 1]
-							.getX2());
+					myCurrentPoint.setControl1(X[k2 - 1].getX1(), Y[k2 - 1].getX1());
+					myCurrentPoint.setControl2(X[k2 - 1].getX2(), Y[k2 - 1].getX2());
 				}
 			} else {
 				c++;
@@ -312,48 +307,45 @@ public class ArcPath implements Shape {
 
 		if (getEndIndex() <= 0) {
 			return;
-		} else if (source != null && source instanceof Transition
-				&& (pathPoints.get(1)).getPointType()) {
+		} else if (source != null && source instanceof Transition && (pathPoints.get(1)).getPointType()) {
 			ArcPathPoint myPoint = pathPoints.get(1);
 			ArcPathPoint myLastPoint = pathPoints.get(0);
-			float distance = (float) getMod(myPoint.getPoint(), myLastPoint
-					.getPoint())
+			double distance = getMod(myPoint.getPoint(), myLastPoint.getPoint())
 					/ Pipe.ARC_CONTROL_POINT_CONSTANT;
-			myPoint.setControl1((float) (myLastPoint.getPoint().x + Math
+			myPoint.setControl1((myLastPoint.getPoint().x + Math
 					.cos(anAngle)
-					* distance), (float) (myLastPoint.getPoint().y + Math
+					* distance), (myLastPoint.getPoint().y + Math
 					.sin(anAngle)
 					* distance));
 
 			myPoint = pathPoints.get(getEndIndex());
-			myPoint.setControl2(getControlPoint(myPoint.getPoint(), myPoint
-					.getControl1(), myPoint.getPoint(), myPoint.getControl1()));
+			myPoint.setControl2(getControlPoint(myPoint.getPoint(), myPoint.getControl1(), myPoint.getPoint(), myPoint.getControl1()));
 		} else if (target != null && source instanceof Place
 				&& (pathPoints.get(getEndIndex())).getPointType()) {
 			ArcPathPoint myPoint = pathPoints.get(getEndIndex());
 			ArcPathPoint myLastPoint = pathPoints.get(getEndIndex() - 1);
-			float distance = (float) getMod(myPoint.getPoint(), myLastPoint
-					.getPoint())
+            double distance = getMod(myPoint.getPoint(), myLastPoint.getPoint())
 					/ Pipe.ARC_CONTROL_POINT_CONSTANT;
-			myPoint.setControl2((float) (myPoint.getPoint().x + Math
+			myPoint.setControl2((myPoint.getPoint().x + Math
 					.cos(anAngle)
-					* distance), (float) (myPoint.getPoint().y + Math
+					* distance), (myPoint.getPoint().y + Math
 					.sin(anAngle)
 					* distance));
 
 			myPoint = pathPoints.get(1);
-			myPoint.setControl1(getControlPoint((pathPoints.get(0)).getPoint(),
-					myPoint.getControl2(), (pathPoints.get(0)).getPoint(),
-					myPoint.getControl2()));
+			myPoint.setControl1(getControlPoint(
+			    (pathPoints.get(0)).getPoint(),myPoint.getControl2(),
+                (pathPoints.get(0)).getPoint(), myPoint.getControl2())
+            );
 		}
 	}
 	
 	public void addPoint(int index, double x, double y, boolean type) {
-		pathPoints.add(index, new ArcPathPoint((float) x, (float) y, type, this));
+		pathPoints.add(index, new ArcPathPoint(x, y, type, this));
 	}
 
 	public void addPoint(double x, double y, boolean type) {
-		pathPoints.add(new ArcPathPoint((float) x, (float) y, type, this));
+		pathPoints.add(new ArcPathPoint(x, y, type, this));
 	}
 	
 
@@ -384,7 +376,7 @@ public class ArcPath implements Shape {
 
 	public void setPointLocation(int index, double x, double y) {
 		if (index < pathPoints.size() && index >= 0) {
-			(pathPoints.get(index)).setPointLocation((float) x, (float) y);
+			(pathPoints.get(index)).setPointLocation(x, y);
 		}
 	}
 
@@ -418,10 +410,10 @@ public class ArcPath implements Shape {
 		return pathPoints.size();
 	}
 
-	public Point2D.Float getPoint(int index) {
+	public Point2D.Double getPoint(int index) {
 		return (pathPoints.get(index)).getPoint();
 	}
-    public Point2D.Float getRealPoint(int index) {
+    public Point2D.Double getRealPoint(int index) {
         return (pathPoints.get(index)).getRealPoint();
     }
 
@@ -583,9 +575,9 @@ public class ArcPath implements Shape {
 	}
 
 	private Cubic[] calcNaturalCubic(int n, int[] x) {
-		float[] gamma = new float[n + 1];
-		float[] delta = new float[n + 1];
-		float[] D = new float[n + 1];
+        double[] gamma = new double[n + 1];
+        double[] delta = new double[n + 1];
+        double[] D = new double[n + 1];
 
 		/*
 		 * We solve the equation [2 1 ] [D[0]] [3(x[1] - x[0]) ] |1 4 1 | |D[1]|
@@ -751,9 +743,9 @@ public class ArcPath implements Shape {
 
 class Cubic {
 
-	float a, b, c, d; /* a + b*u + c*u^2 + d*u^3 */
+	double a, b, c, d; /* a + b*u + c*u^2 + d*u^3 */
 
-	public Cubic(float _a, float _b, float _c, float _d) {
+	public Cubic(double _a, double _b, double _c, double _d) {
 		a = _a;
 		b = _b;
 		c = _c;
@@ -761,17 +753,17 @@ class Cubic {
 	}
 
 	// Return first control point coordinate (calculated from coefficients)
-	public float getX1() {
+	public double getX1() {
 		return ((b + 3 * a) / 3);
 	}
 
 	// Return second control point coordinate (calculated from coefficients)
-	public float getX2() {
+	public double getX2() {
 		return ((c + 2 * b + 3 * a) / 3);
 	}
 
 	/** evaluate cubic */
-	public float eval(float u) {
+	public double eval(double u) {
 		return (((d * u) + c) * u + b) * u + a;
 	}
