+++ src/pipe/gui/graphicElements/ArcPathPoint.java
@@ -33,11 +33,11 @@ public class ArcPathPoint extends PetriNetObject {
 	private final int DELTA = 10;
 
     private ArcPath myArcPath;
-	private Point2D.Float point = new Point2D.Float();
-	private Point2D.Float realPoint = new Point2D.Float();
+	private Point2D.Double point = new Point2D.Double();
+	private Point2D.Double realPoint = new Point2D.Double();
 
-	private Point2D.Float control1 = new Point2D.Float();
-	private Point2D.Float control2 = new Point2D.Float();
+	private Point2D.Double control1 = new Point2D.Double();
+	private Point2D.Double control2 = new Point2D.Double();
 
 	private boolean pointType; // STRAIGHT or CURVED
 
@@ -72,7 +72,7 @@ public class ArcPathPoint extends PetriNetObject {
 		this(point.x, point.y, _pointType, a);
 	}
 
-	public Point2D.Float getPoint() {
+	public Point2D.Double getPoint() {
 		return point;
 	}
 
@@ -112,7 +112,7 @@ public class ArcPathPoint extends PetriNetObject {
 		myArcPath.setPointVisibilityLock(lock);
 	}
 
-	public double getAngle(Point2D.Float p2) {
+	public double getAngle(Point2D.Double p2) {
 		double angle;
 
 		if (point.y <= p2.y) {
@@ -198,30 +198,30 @@ public class ArcPathPoint extends PetriNetObject {
 		return (i > 0 && i != myArcPath.getNumPoints() - 1);
 	}
 
-	public Point2D.Float getControl1() {
+	public Point2D.Double getControl1() {
 		return control1;
 	}
 
-	public Point2D.Float getControl2() {
+	public Point2D.Double getControl2() {
 		return control2;
 	}
 
-	public void setControl1(float _x, float _y) {
+	public void setControl1(double _x, double _y) {
 		control1.x = _x;
 		control1.y = _y;
 	}
 
-	public void setControl2(float _x, float _y) {
+	public void setControl2(double _x, double _y) {
 		control2.x = _x;
 		control2.y = _y;
 	}
 
-	public void setControl1(Point2D.Float p) {
+	public void setControl1(Point2D.Double p) {
 		control1.x = p.x;
 		control1.y = p.y;
 	}
 
-	public void setControl2(Point2D.Float p) {
+	public void setControl2(Point2D.Double p) {
 		control2.x = p.x;
 		control2.y = p.y;
 	}
@@ -261,18 +261,18 @@ public class ArcPathPoint extends PetriNetObject {
 		} else {
 			SIZE = 3;
 		}
-		float x = Zoomer.getZoomedValue(getRealPoint().x, zoom);
-		float y = Zoomer.getZoomedValue(getRealPoint().y, zoom);
+		double x = Zoomer.getZoomedValue(getRealPoint().x, zoom);
+		double y = Zoomer.getZoomedValue(getRealPoint().y, zoom);
 		point.setLocation(x, y);
 		setBounds((int) x - SIZE, (int) y - SIZE, 2 * SIZE + SIZE_OFFSET, 2
 				* SIZE + SIZE_OFFSET);
 	}
 
-	public Point2D.Float getRealPoint() {
+	public Point2D.Double getRealPoint() {
 		return realPoint;
 	}
 
-	public void setRealPoint(Point2D.Float realPoint) {
+	public void setRealPoint(Point2D.Double realPoint) {
 		this.realPoint = realPoint;
 	}
